Moved code out of src to work with our pipeline
