+++ NLog.Targets.ElasticSearch.Tests/TestOutputTextWriter.cs
