+++ NLog.Targets.ElasticSearch.sln
