+++ NLog.Targets.ElasticSearch/ExceptionExtensions.cs
