+++ NLog.Targets.ElasticSearch/ExpandoObjectExtensions.cs
