+++ NLog.Targets.ElasticSearch/NLog.Targets.ElasticSearch.snk
