+++ NLog.Targets.ElasticSearch/StringExtensions.cs
