--- src/NLog.Targets.ElasticSearch.Tests/App.config
@@ -1,24 +1,0 @@
-<?xml version="1.0" encoding="utf-8"?>
-<configuration>
-  <configSections>
-    <section name="nlog" type="NLog.Config.ConfigSectionHandler, NLog"/>
-  </configSections>
-  <nlog xmlns="http://www.nlog-project.org/schemas/NLog.xsd" xmlns:xsi="http://www.w3.org/2001/XMLSchema-instance">
-    <extensions>
-      <add assembly="NLog.Targets.ElasticSearch"/>
-    </extensions>
-    <targets>
-      <target name="elastic" xsi:type="BufferingWrapper" flushTimeout="5000">
-        <target xsi:type="ElasticSearch" layout="${logger} | ${threadid} | ${message}" includeAllProperties="true" requireAuth="true" username="user" password="pwd">
-          <field name="user" layout="${windows-identity:userName=True:domain=False}"/>
-          <field name="host" layout="${machinename}"/>
-          <field name="number" layout="1" layoutType="System.Int32"/>
-          <field name="test" layout="${event-properties:item=test}"/>
-        </target>
-      </target>
-    </targets>
-    <rules>
-      <logger name="*" minlevel="Info" writeTo="elastic"/>
-    </rules>
-  </nlog>
-<startup><supportedRuntime version="v4.0" sku=".NETFramework,Version=v4.7.2"/></startup></configuration>
